requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bidict==0.23.1
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
deep-translator==1.11.4
distro==1.9.0
dnspython==2.7.0
eventlet==0.40.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==2.0.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fpdf==1.7.2
fsspec==2025.5.1
future==1.0.0
greenlet==3.2.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
openai==1.82.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
pillow==10.4.0
proglog==0.1.11
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==0.19.0
python-engineio==4.12.1
python-socketio==5.13.0
pytube==12.1.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rfc3986==1.5.0
setuptools==78.1.0
simple-websocket==1.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
SpeechRecognition==3.8.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==10.0
Werkzeug==2.0.1
wsproto==1.2.0
